Keep shared prepared RSS helper portable
