+++ TreeDB/collections/column_vector_graph_shared_prepared_rss_windows_test.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package collections
+
+func processMaxRSSBytes1735() uint64 {
+	return 0
+}
