+++ TreeDB/collections/column_vector_graph_shared_prepared_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"runtime"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 )
@@ -280,18 +279,4 @@ func runtimeHeapAllocBytes1735() uint64 {
 	var mem runtime.MemStats
 	runtime.ReadMemStats(&mem)
 	return mem.HeapAlloc
-}
-
-func processMaxRSSBytes1735() uint64 {
-	var usage syscall.Rusage
-	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &usage); err != nil || usage.Maxrss <= 0 {
-		return 0
-	}
-	maxRSS := uint64(usage.Maxrss)
-	switch runtime.GOOS {
-	case "linux", "freebsd", "openbsd", "netbsd":
-		return maxRSS * 1024
-	default:
-		return maxRSS
-	}
 }
